% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lterpalettefinder-package.R
\docType{package}
\name{lterpalettefinder-package}
\alias{lterpalettefinder}
\alias{lterpalettefinder-package}
\title{lterpalettefinder: Extract Color Palettes from Photos and Pick Official LTER Palettes}
\description{
Allows identification of palettes derived from LTER (Long Term Ecological Research) photographs based on user criteria. Also facilitates extraction of palettes from users' photos directly.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/lter/lterpalettefinder/issues}
}

}
\author{
\strong{Maintainer}: Nicholas J Lyon \email{njlyon@alumni.iastate.edu} (\href{https://orcid.org/0000-0003-3905-1078}{ORCID})

Other contributors:
\itemize{
  \item Gabe De La Rosa \email{delarosa@nceas.ucsb.edu} [contributor]
  \item Julien Brun \email{brun@nceas.ucsb.edu} (\href{https://orcid.org/0000-0002-7751-6238}{ORCID}) [contributor]
  \item  National Science Foundation (NSF 1929393, 09/01/2019 - 08/31/2024) [funder]
}

}
\keyword{internal}
