% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_check.R
\name{palette_check}
\alias{palette_check}
\title{Check if a palette is using the correct format}
\usage{
palette_check(palette)
}
\arguments{
\item{palette}{Vector of hexadecimal codes returned by `palette_extract()`, `..._sort()`, or `..._subsample()`}
}
\value{
An error message or nothing
}
\description{
Accepts the hexadecimal code vector and test if it is of the correct format
}
\examples{
# Check for misformatted hexcodes
\dontrun{
palette_check(c("#8e847a", "#9fc7f2g3"))
}
}
