% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_demo.R
\name{palette_demo}
\alias{palette_demo}
\title{Demonstrate Extracted Palette with HEX Labels}
\usage{
palette_demo(
  palette,
  export = FALSE,
  export_name = "my_palette",
  export_path = getwd()
)
}
\arguments{
\item{palette}{One-column dataframe returned by `palette_extract()`}

\item{export}{Logical of whether or not to export the demo plot}

\item{export_name}{Character string of what to name exported plot}

\item{export_path}{File path to save exported plot (defaults to working directory)}
}
\value{
A plot of the retrieved colors labeled with their HEX codes
}
\description{
Accepts the hexadecimal code dataframe returned by `palette_extract()` and creates a simple plot of all returned colors labeled with their HEX codes. This will facilitate (hopefully) your selection of which of the 25 colors you would like to use in a given context.
}
\examples{
## This example is commented out because that's not a real file path or PNG file.
## This example will work if you change the first line to point to an existing .PNG

## Extract colors from a supplied image
# my_colors <- palette_extract(image = file.path("Path", "to", "picture.png"),
#  progress_bar = TRUE)

## Create a plot to evaluate the HEX codes and facilitate selection of which you want
# palette_demo(palette = my_colors)

}
