% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_extract.R
\name{palette_extract}
\alias{palette_extract}
\title{Extract Hexadecimal Codes from an Image}
\usage{
palette_extract(image, progress_bar = TRUE)
}
\arguments{
\item{image}{Name/path to PNG, JPEG, or TIFF file from which to extract colors}

\item{progress_bar}{Logical (TRUE / FALSE) indicating whether a progress bar is desired}
}
\value{
A dataframe of a single column ("hex_code") containing all hexadecimal codes remaining after extraction and removal of 'dark' and 'similar' colors.
}
\description{
Retrieves hexadecimal codes for the colors in an image file. Currently only PNG, JPEG, and TIFF files are supported. The function automatically removes dark colors and removes 'similar' colors to yield 25 colors from which you can select the subset that works best for your visualization needs. Note that photos that are very dark may return few viable colors.
}
\examples{
# Extract colors from a supplied image
my_colors <- palette_extract(image = system.file("extdata", "lyon-fire.png",
package = "lterpalettefinder"))

# Look at the returned hexadecimal codes
       
# Plot that result
palette_demo(palette = my_colors)

}
