% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_finder.R
\name{palette_finder}
\alias{palette_finder}
\title{Find an LTER Palette}
\usage{
palette_finder(site = "all", name = "all", type = "all", length = "all")
}
\arguments{
\item{site}{Vector of three-letter LTER site abbreviations for which to return palettes}

\item{name}{Vector of palette names (if known) for which to return palettes}

\item{type}{Vector of palette types (e.g., qualitative, sequential, etc.) for which to return palettes}

\item{length}{Vector of acceptable palette lengths (i.e., how many colors are needed)}
}
\value{
If more than one palette is a dataframe, if exactly one palette is a vector
}
\description{
From a dataframe of all possible palettes (updated periodically so check back!) specify the characteristics of the palette you want and retrieve the palettes that match those criteria. Can specify by number of colors in the palette, type of palette (e.g., qualitative, sequential, etc.), or which LTER site the palette came from.
}
\examples{
# Look at all palette options by calling the function without specifying arguments
lterpalettefinder::palette_finder()

# What if our query returns NO options?
lter_palette_finder(length = 1)

# What if our query returns MULTIPLE options?
lter_palette_finder(length = 5, site = "HBR")

# What if our query returns JUST ONE option? (this is desirable)
lter_palette_finder(site = "AND")

}
