% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_check.R
\name{palette_check}
\alias{palette_check}
\title{Check Hexadecimal Code Formatting}
\usage{
palette_check(palette)
}
\arguments{
\item{palette}{(character) Vector of hexadecimal codes returned by `palette_extract()`, `..._sort()`, or `..._subsample()`}
}
\value{
An error message or nothing
}
\description{
Accepts the hexadecimal code vector and tests if it is formatted correctly
}
\examples{
# Check for misformatted hexcodes
palette_check(palette = c("#8e847a", "#9fc7f2"))
}
