% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_extract.R
\name{palette_extract}
\alias{palette_extract}
\title{Extract Hexadecimal Codes from an Image}
\usage{
palette_extract(image, progress_bar = TRUE)
}
\arguments{
\item{image}{Name of the .PNG file to extract colors from (only .PNGs are supported at this time)}

\item{progress_bar}{Logical (TRUE / FALSE) indicating whether printing a progress bar in the console as the function extracts colors is desired}
}
\value{
A dataframe of a single column ("hex_code") containing all hexadecimal codes remaining after extraction and removal of 'dark' and 'similar' colors.
}
\description{
Retrieves hexadecimal codes for the colors in an image file. Currently only .PNG files are supported. The function automatically removes dark colors and removes 'similar' colors to yield 25 colors from which you can select the subset that works best for your visualization needs. Note that photos that are very dark may return few viable colors.
}
\examples{
## This example is commented out because that's not a real file path or PNG file.
## This example will work if you change the first line to point to an existing .PNG

## Extract colors from a supplied image
# my_colors <- palette_extract(image = file.path("Path", "to", "picture.png"),
#  progress_bar = TRUE)

## Look at the returned HEX codes
# my_colors
       

}
